+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v0.15.19 (??.??.????)
  * new Wrapper function tx_rnbase_util_Strings::inList
  * new Wrapper function tx_rnbase_util_Network::getUrl
  * new Wrapper function tx_rnbase_parameters::getPostAndGetParametersMerged for _GPmerged
+ * new Wrapper function tx_rnbase_parameters::getPostAndGetParameter for _GP
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall